Move ingest source manager to persistence module
