+++ admin/pom.xml
@@ -82,11 +82,6 @@
             <artifactId>jackson-databind</artifactId>
         </dependency>
 
-        <dependency>
-            <groupId>gov.loc</groupId>
-            <artifactId>bagit</artifactId>
-        </dependency>
-
         <!-- XML -->
         <dependency>
             <groupId>jaxen</groupId>
