+++ persistence/src/main/java/edu/unc/lib/dl/persist/services/ingest/IngestSourceManager.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.unc.lib.dl.admin.collect;
+package edu.unc.lib.dl.persist.services.ingest;
 
 import java.io.File;
 import java.io.IOException;
