+++ persistence/src/test/java/edu/unc/lib/dl/persist/services/ingest/IngestSourceManagerTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.unc.lib.dl.admin.collect;
+package edu.unc.lib.dl.persist.services.ingest;
 
 import static org.mockito.MockitoAnnotations.initMocks;
